Added shader double support

+++ shaders/screen.frag
@@ -4,29 +4,29 @@
 
 layout(set = 0, binding = 0) uniform Variables
 {
-	vec3 transformation;
+	dvec3 transformation;
 } variables;
 
 layout(location = 0) out vec4 pixelColor;
 
-const vec2 resolution = vec2(1600, 900);
-const float aspect = resolution.x / resolution.y;
+const dvec2 resolution = vec2(1600, 900);
+const double aspect = resolution.x / resolution.y;
 
 int maxIterations = 90;
 //float zoom = 1.0;
 //vec2 center = vec2(-0.5, 0.0);
 
-float Manderbrot(vec2 uv)
+double Manderbrot(dvec2 uv)
 {
-	float x = (uv.x - 0.5) * 2.0 * aspect / variables.transformation.z + variables.transformation.x;
-	float y = (uv.y - 0.5) * 2.0 / variables.transformation.z + variables.transformation.y;
-	vec2 c = vec2(x, y);
-	vec2 z = vec2(0.0);
+	double x = (uv.x - 0.5) * 2.0 * aspect / variables.transformation.z + variables.transformation.x;
+	double y = (uv.y - 0.5) * 2.0 / variables.transformation.z + variables.transformation.y;
+	dvec2 c = dvec2(x, y);
+	dvec2 z = dvec2(0.0);
 
 	int iterations;
 	for (iterations = 0; iterations < maxIterations; iterations++)
 	{
-		float tmpX = (z.x * z.x - z.y * z.y) + c.x;
+		double tmpX = (z.x * z.x - z.y * z.y) + c.x;
 		z.y = 2.0 * z.x * z.y + c.y;
 		z.x = tmpX;
 
@@ -36,14 +36,14 @@ float Manderbrot(vec2 uv)
 		}
 	}
 
-	return (float(iterations) / float(maxIterations));
+	return (double(iterations) / double(maxIterations));
 }
 
 void main()
 {
 	//pixelColor = vec4(gl_FragCoord.x / 800.0, gl_FragCoord.y / 800.0, 1.0, 1.0);
 
-	vec2 uv = gl_FragCoord.xy / resolution.xy;
-	float result = Manderbrot(uv);
+	dvec2 uv = gl_FragCoord.xy / resolution.xy;
+	double result = Manderbrot(uv);
 	pixelColor = vec4(vec3(result), 1.0);
 }
